add bfly function (work not correct)

+++ libft/ft_lstsize.c
@@ -6,15 +6,15 @@
 /*   By: ggevorgi <sp1tak.gg@gmail.com>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/02 20:11:15 by ggevorgi          #+#    #+#             */
-/*   Updated: 2025/02/11 12:22:09 by ggevorgi         ###   ########.fr       */
+/*   Updated: 2025/02/13 13:18:47 by ggevorgi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_lstsize(t_stack_node *lst)
+size_t	ft_lstsize(t_stack_node *lst)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while (lst)
